requirements.txt
fastapi
uvicorn
transformers
torch
langchain
fitz
pytesseract
pillow
chromadb
sentence-transformers
PyMuPDF
python-multipart
kagglehub

